Fix hostapd_deinit_wpa() in init error cases
With driver wrappers that implement set_privacy() or set_generic_elem(),
it was possible to hit a NULL pointer dereference in an error case where
interface setup failed and the network configuration used WPA/WPA2.

Fix this by skipping the driver operations in case the driver interface
is not initialized.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>